backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
slowapi>=0.1.9
groq>=0.9.0
google-genai>=1.0.0
chromadb>=0.5.5
sentence-transformers>=3.0.1
pypdf>=4.2.0
tiktoken>=0.7.0
